requirements.txt
discord.py
rich
python-dotenv
aiohttp
watchdog